Refactoring backend
