+++ java-tools/OsmAndServer/src/main/java/net/osmand/server/api/searchtest/dto/RunStatus.java
@@ -1,7 +1,6 @@
 package net.osmand.server.api.searchtest.dto;
 
 import net.osmand.server.api.searchtest.entity.Run;
-import net.osmand.server.api.searchtest.entity.TestCase;
 
 import java.util.Map;
 
@@ -12,4 +11,5 @@ public record RunStatus(
 		long duration,
 		double averagePlace,
 		long found,
-		Map<String, Number> distanceHistogram) {}
+		Map<String, Number> distanceHistogram,
+		TestCaseStatus generatedChart) {}
