+++ java-tools/OsmAndServer/src/main/java/net/osmand/server/api/searchtest/dto/TestCaseStatus.java
@@ -2,9 +2,7 @@ package net.osmand.server.api.searchtest.dto;
 
 import net.osmand.server.api.searchtest.entity.TestCase;
 
-import java.util.Map;
-
-public record TestStatus(
+public record TestCaseStatus(
 		TestCase.Status status,
 		long processed,
 		long failed,
